Some changes
-Form file moved to root
-HTTP request transfered from javascript to php
-If a mail is used from both intern and extern list, the 2nd one is suppressed

+++ index.php
@@ -8,7 +8,9 @@ if(isset($_GET['action']) && $_GET['action']=='login'){
   $myUser=$setUser->getUserbyMail($_POST['email']);
   if (is_a ($myUser,'UserEntity')){
   $_SESSION['user']=$myUser;}
-  else {$error='Erreur de connexion';}
+  else {$_SESSION['error']='Erreur de connexion';}
+  header('Location: index.php');
+  exit();
   }
 
   if(isset($_GET['action']) && $_GET['action']=='logout'){
@@ -34,13 +36,13 @@ if(isset($_GET['action']) && $_GET['action']=='login'){
 <body>
   <div class="container">
     <header class="navbar">
-      <div><a class="btn btn-success" href="inc/form.php">Create an account</a></div>
+      <div><a class="btn btn-success" href="form.php">Create an account</a></div>
       <div>
         <? if(isset($_SESSION['user'])){
           ?>
           <p>Bienvenue <?= $_SESSION['user']->getFirst_Name() ?> !</p>
           <p><a class="btn btn-success" href="?action=logout">Log out</a></p>
-       <?php }else{ ?>
+       <?php } else { ?>
       <form action="?action=login" method="post">
         <label for="">Email</label>
         <input type="text" name="email">
@@ -48,28 +50,50 @@ if(isset($_GET['action']) && $_GET['action']=='login'){
        </form> <?php } ?></div>
       <div></div>
     </header>
+       <?php if(!is_null($_SESSION['error'])){ ?><div class="alert alert-danger"><?php echo $_SESSION['error'];$_SESSION['error']=null; ?></div> <?php } ?>
 
 
 <?php 
-$users = $setUser->getAllUsers();
+
 
 /* ?><pre><?
   var_dump($user);
 ?></pre> */
 ?> <h2>Users from Beewake</h2><?php
+if($users){
+  
 foreach($users as $user){ ?>
 <div class="row" class="user">
 <img class="col-4" src="<?php echo $user->getAvatar(); ?>" alt="Card image cap">
  <div class="col-8">
  <h5 class="card-title"><?= strtoupper($user->getFirst_name()).' '.strtoupper($user->getLast_name()) ?></h5>
   <p class="card-text"><?= $user->getEmail() ?></p>
-  <a class="btn btn-success" href="inc/form.php?id=<?= $user->getId() ?>">Edit profil</a>
+  <a class="btn btn-success" href="form.php?id=<?= $user->getId() ?>">Edit profil</a>
   <a class="btn btn-success" href="?action=delete&id=<?= $user->getId() ?>">Delete profil</a>
   </div>
   </div>
   <hr />
-<?php } ?>
-<div id="content"></div>
+<?php
+
+
+} }
+
+?> <h2>Users from partners</h2>
+foreach($_SESSION['users']['data'] as $guestUser){
+
+    if(!$setUser->getUserByMail($guestUser['email'] ) ) {
+      ?>
+      <div class="row" class="user">
+      <img class="col-4" src="<?php echo $guestUser['avatar']; ?>" alt="Card image cap">
+       <div class="col-8">
+       <h5 class="card-title"><?= strtoupper($guestUser['first_name']).' '.strtoupper($guestUser['last_name']) ?></h5>
+        <p class="card-text"><?= $guestUser['email'] ?></p>
+        </div>
+        </div>
+        <hr />
+   <?php } 
+ } ?>
+
 </div>
 </body>
 
@@ -79,5 +103,5 @@ integrity="sha256-FgpCb/KJQlLNfOu91ta32o/NMZxltwRo8QtmkMRdAu8="
 crossorigin="anonymous"></script>
 <script src="https://cdn.jsdelivr.net/npm/js-cookie@2/src/js.cookie.min.js"></script>
 <script src="js/User.js"></script>
-<script src="js/index.js"></script>
+<!-- <script src="js/index.js"></script> -->
 </html>
